requirements.txt
Flask
pyjwt
Flask-Script
mysql-connector